fix: Fix syntax issues

+++ star-dash/star-dash/GameEngine/Components/BuffComponent.swift
@@ -1,7 +1,7 @@
 import Foundation
 
 class BuffComponent: Component {
-    var speedMultiplier: Float = 1
+    var speedMultiplier: CGFloat = 1
 
     override init(id: ComponentId, entityId: EntityId) {
         super.init(id: id, entityId: entityId)
